src/timelinextract/requirements.txt
alabaster==0.7.16
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
babel==2.16.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.2
defusedxml==0.7.1
distro==1.9.0
dj-rest-auth==7.0.1
Django==5.1.5
django-allauth==65.5.0
django-cors-headers==4.7.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
docutils==0.20.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
imagesize==1.4.1
Jinja2==3.1.5
jiter==0.8.2
MarkupSafe==3.0.2
oauthlib==3.2.2
openai==1.59.8
packaging==24.2
pdfservices-sdk==4.1.0
pycparser==2.22
pydantic==2.10.5
pydantic_core==2.27.2
Pygments==2.19.1
PyJWT==2.9.0
PyPDF2==3.0.1
python-dotenv==1.0.1
python3-openid==3.2.0
requests==2.31.0
requests-oauthlib==2.0.0
sniffio==1.3.1
snowballstemmer==2.2.0
social-auth-app-django==5.4.3
social-auth-core==4.5.6
Sphinx==7.3.7
sphinx-rtd-theme==2.0.0
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
sqlparse==0.5.3
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.3.0
